+++ scripts/macros.js
@@ -201,6 +201,7 @@ export {mageHand, mageHandEffect} from './macros/2024/spells/mageHand.js';
 export {guidance} from './macros/2024/spells/guidance.js';
 export {sproutFoliage} from './macros/2024/spells/sproutFoliage.js';
 export {forestGuard, forestGuardTurnEffect, forestGuardEffect} from './macros/2024/spells/forestGuard.js';
+export {cacophonicShield, cacophonicShieldSourceEffect} from './macros/2024/spells/cacophonicShield.js';
 // Items
 export {healersKit} from './macros/2024/items/healersKit.js';
 export {unarmedStrike} from './macros/2024/items/misc/unarmedStrike.js';